+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/)
 
+## [0.1.0-beta.17] - 2021-02-06
+
+### Fixed
+
+- An issue where too many notifications where created on request accept
+
+### Added
+
+- More details to notifications
+- Discord notification via PM to the user on SRP request approval or rejection, if
+  [AA-Discordbot](https://github.com/pvyParts/allianceauth-discordbot) is installed
+
 
 ## [0.1.0-beta.16] - 2021-02-06